Move build_services
