+++ src/compiler/typescript_to_rust/azle_generate/src/ts_ast/ts_ast/mod.rs
@@ -6,7 +6,6 @@ use swc_ecma_parser::{lexer::Lexer, Parser, StringInput, Syntax, TsConfig};
 use crate::ts_ast::{source_map::SourceMapped, AzleProgram};
 
 mod build_vm_value_conversion;
-mod cross_canister_calls;
 
 pub struct TsAst {
     pub azle_programs: Vec<AzleProgram>,
